Refactor table initialization logic for SQLite and PostgreSQL
Moved SQLite and PostgreSQL table creation queries into separate files (`sqliteTables.js` and `postgresTables.js`) for better modularity and maintainability. Updated `DatabaseStrategy` to use these modular queries, improving code readability and reducing duplication.
