+++ queries/postgresTables.js
@@ -1,0 +1,43 @@
+export const createPostgresTables = `
+CREATE TABLE IF NOT EXISTS users (
+    id SERIAL PRIMARY KEY,
+    first_name VARCHAR(50),
+    last_name VARCHAR(50),
+    email VARCHAR(100) UNIQUE,
+    apikey VARCHAR(50) UNIQUE,
+    username VARCHAR(100) UNIQUE,
+    file_permissions TEXT[],
+    created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
+    updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
+);
+
+CREATE TABLE IF NOT EXISTS images (
+    id SERIAL PRIMARY KEY,
+    user_id INT NOT NULL,
+    filename VARCHAR(255) NOT NULL,
+    path VARCHAR(500) NOT NULL,
+    file_type VARCHAR(100),
+    size INT,
+    original_filename VARCHAR(255),
+    original_file_type VARCHAR(100),
+    original_size INT,
+    upload_date TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
+    tags TEXT[],
+    metadata JSONB,
+    CONSTRAINT fk_user FOREIGN KEY (user_id) REFERENCES users (id)
+);
+
+CREATE OR REPLACE FUNCTION update_updated_at()
+RETURNS TRIGGER AS $$
+BEGIN
+    NEW.updated_at = CURRENT_TIMESTAMP;
+    RETURN NEW;
+END;
+$$ LANGUAGE plpgsql;
+
+DROP TRIGGER IF EXISTS update_users_updated_at ON users;
+CREATE TRIGGER update_users_updated_at
+BEFORE UPDATE ON users
+FOR EACH ROW
+EXECUTE FUNCTION update_updated_at();
+`;
